Refactor file system save and retrieve logic into FileSystemMultimediaFileManager
